stage 1
- create server object in module, and move feature registration outside of GrizzlyLanguageServer init
- split up tests to unit and e2e
- keep LspFixture alive between test functions to speed up tests, requires a bit more restoration of state between tests
