+++ grizzly-ls/tests/unit/test_progress.py
@@ -1,8 +1,9 @@
 from unittest.mock import ANY
 
 from grizzly_ls.progress import Progress
+from pytest_mock import MockerFixture
 
-from tests.fixtures import LspFixture, MockerFixture
+from tests.fixtures import LspFixture
 
 
 def test_progress(lsp_fixture: LspFixture, mocker: MockerFixture) -> None:
